requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Machine Learning & AI
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
transformers>=4.30.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# Exchange APIs
python-binance>=1.0.19
pybit>=2.4.1
kucoin-python>=1.0.11
ccxt>=4.0.0

# WebSocket & Async
websockets>=11.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.16.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0

# Web Framework (for GUI/API)
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
plotly>=5.15.0

# Mobile App (Kivy)
kivy>=2.2.0
kivymd>=1.1.1

# Utilities
loguru>=0.7.0
python-json-logger>=2.0.7
colorama>=0.4.6
tqdm>=4.65.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Monitoring & Metrics
prometheus-client>=0.17.0
psutil>=5.9.0
